+++ ui/src/app/core/accessories/types/filtermaintenance/filtermaintenance.component.ts
@@ -1,9 +1,12 @@
 import { NgClass } from '@angular/common'
-import { Component, Input } from '@angular/core'
+import { Component, inject, Input } from '@angular/core'
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap'
 import { TranslatePipe } from '@ngx-translate/core'
 import { InlineSVGModule } from 'ng-inline-svg-2'
 
 import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
+import { FilterMaintenanceManageComponent } from '@/app/core/accessories/types/filtermaintenance/filtermaintenance.manage.component'
+import { LongClickDirective } from '@/app/core/directives/longclick.directive'
 
 @Component({
   selector: 'app-filtermaintenance',
@@ -13,10 +16,21 @@ import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
     InlineSVGModule,
     NgClass,
     TranslatePipe,
+    LongClickDirective,
   ],
 })
 export class FilterMaintenanceComponent {
+  private $modal = inject(NgbModal)
+
   @Input() public service: ServiceTypeX
 
   constructor() {}
+
+  onClick() {
+    const ref = this.$modal.open(FilterMaintenanceManageComponent, {
+      size: 'md',
+      backdrop: 'static',
+    })
+    ref.componentInstance.service = this.service
+  }
 }