+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - battery - style as green when charging
 - thermostat - open control modal on long click
 - show release notes when npm and gh release match
+- add control modal for filter to reset filter life
 
 ### Homebridge Dependencies